Refactor reaper to allow for updating watching resources
